drm/i915: Fixed Compilation Warnings
BZ: 145899

After porting 3.10 compiler warnings have been fixed.
A subsequent patch will be sent to resolve all
remaining warnings.

Reviewed-on: http://android.intel.com:8080/139215
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Tested-by: cactus <cactus@intel.com>

+++ drivers/gpu/drm/i915/i915_drv.h
@@ -2496,7 +2496,7 @@ void vlv_force_wake_put(struct drm_i915_private *dev_priv, int fw_engine);
 /* intel_sideband.c */
 u32 vlv_punit_read(struct drm_i915_private *dev_priv, u8 addr);
 void vlv_punit_write(struct drm_i915_private *dev_priv, u8 addr, u32 val);
-u32 vlv_nc_read(struct drm_i915_private *dev_priv, u8 addr);
+u32 vlv_nc_read(struct drm_i915_private *dev_priv, u16 addr);
 u32 vlv_gpio_nc_read(struct drm_i915_private *dev_priv, u32 reg);
 void vlv_gpio_nc_write(struct drm_i915_private *dev_priv, u32 reg, u32 val);
 u32 vlv_cck_read(struct drm_i915_private *dev_priv, u32 reg);
